Add initial impl of tree traversal
This is to map Python code definitions to filesystem names. `test.py`
contains code to check and `test1.py` is used to check `test.py` for
(its own) correctness.

Perhaps `recurse_filter` should use dicts internally instead? It will
probably be a bit longer but it makes the code to construct the '.d' and
'.c' "directories" easier.